+++ src/compression_fs_extension.cpp
@@ -5,6 +5,7 @@
 #include "compression_fs_extension.hpp"
 #include "lz4_file_system.hpp"
 #include "snappy_file_system.hpp"
+#include "xz_file_system.hpp"
 
 #include "duckdb.hpp"
 #include "duckdb/main/extension/extension_loader.hpp"
@@ -12,12 +13,14 @@
 namespace duckdb {
 
 static void LoadInternal(ExtensionLoader &loader) {
-	loader.SetDescription("Compression filesystems that DuckDB does not ship natively (brotli, bzip2, lz4, snappy)");
+	loader.SetDescription(
+	    "Compression filesystems that DuckDB does not ship natively (brotli, bzip2, lz4, snappy, xz)");
 	auto &fs = loader.GetDatabaseInstance().GetFileSystem();
 	fs.RegisterCompressionFilesystem(make_uniq<BrotliFileSystem>());
 	fs.RegisterCompressionFilesystem(make_uniq<Bzip2FileSystem>());
 	fs.RegisterCompressionFilesystem(make_uniq<Lz4FileSystem>());
 	fs.RegisterCompressionFilesystem(make_uniq<SnappyFileSystem>());
+	fs.RegisterCompressionFilesystem(make_uniq<XzFileSystem>());
 }
 
 void CompressionFsExtension::Load(ExtensionLoader &loader) {